// web/projects/frontend/src/app/store/push-subscription/models/push-subscription-state.model.ts
/**
 * @license
 * Copyright Dutch Cask. All Rights Reserved.
 *
 * Use of this source code is governed by an MIT-style license that
 * can be found in the LICENSE file at https://dutch-cask.nl/license
 */

export interface PushSubscriptonState {
  isPushAllowed: boolean;
  isPushEnabled: boolean;
  isSubscribed: boolean;
}

export interface SubscriptionState {
  state: PushSubscriptonState;
}